// src/components/Models/Enumerations.ts
export enum FormInputTypes {
  Select = 1,
  Text = 2,
  Checkbox = 3,
  DatePicker = 4,
  Hidden = 5,
  Password = 6,
  Email = 7,
  Number = 8,
  TextArea = 9,
  Switch = 10,
  Date = 11,
}

export enum ValidationType {
  required = 1,
  isEmail = 2,
  minLength = 3,
  maxLength = 4,
  isTrue = 5,
  passwordMatch = 6,
}



export enum FilterTypes {
  Select = 1,
  Text = 2,
  Checkbox = 3,
  DatePicker = 4,
  Number,
}



export enum TransactionType {
  Credit = 1,
  Debit = 2,
}
export enum TransactionFor {
  PurchaseOrder = 1,
  SaleOrder = 2,
}